improvements

+++ puppeteer-test/src/pup-testing.js
@@ -22,14 +22,19 @@ const observe = [
 ];
 
 async function pupTesting(page, host) {
-    fs.readdir(`screenshots-${host}`, (err, files) => {
-        if (err) throw err;
-        for (const file of files) {
-            fs.unlink(path.join(`screenshots-${host}`, file), err => {
-                if (err) throw err;
-            });
-        }
-    });
+    const dir = `screenshots-${host}`;
+    if (!fs.existsSync(dir)){
+        fs.mkdirSync(dir);
+    } else {
+        fs.readdir(dir, (err, files) => {
+            if (err) throw err;
+            for (const file of files) {
+                fs.unlink(path.join(`screenshots-${host}`, file), err => {
+                    if (err) throw err;
+                });
+            }
+        });
+    }
 
     // register events listeners
     const client = await page.target().createCDPSession();
@@ -49,201 +54,201 @@ async function pupTesting(page, host) {
     const complete = [];
 
     complete.push(Date.now());
-    await page.goto(`https://${host}.picturedesk.ch`, {waitUntil: 'networkidle2'});
-    complete.push(Date.now()-complete.pop());
+    await page.goto(`https://${host}.picturedesk.ch`, {waitUntil: 'networkidle0'});
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // ÜBER UNS
     // Kontakt
     complete.push(Date.now());
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.meta-nav a:nth-child(1)')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Über Uns
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(1) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Qualität
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(2) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Kooperation
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) ul li:nth-child(1) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Gesundheitsprogramme
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) ul li:nth-child(2) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Integrierte Versorgung
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(3) ul li:nth-child(3) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Kontakt
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(4) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Kongresse
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(5) ul li:nth-child(1) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // 2019, Traditionell vor Headless
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(2) a');
     await page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(2) ul li:nth-child(5) ul li:nth-child(2) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // SPITAL / KLINIK
     complete.push(Date.now());
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(3) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // APOTHEKEN
     complete.push(Date.now());
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(4) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Ärzte / Zahnärzte
     // Ärzte / Zahnärzte
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(1) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // TARMED
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(2) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Ärzte
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(3) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Ärztenetzwerke
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(4) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Hausärzte
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(5) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // Zweitmeinungsarzt
     complete.push(Date.now());
     await page.click('.navbar-nav > li:nth-child(5) a');
     await Promise.all([
-        page.waitForNavigation({waitUntil: 'networkidle2'}),
+        page.waitForNavigation({waitUntil: 'networkidle0'}),
         page.click('.navbar-nav > li:nth-child(5) ul li:nth-child(6) a')
     ]);
-    complete.push(Date.now()-complete.pop());
     await page.screenshot({path: `screenshots-${host}/screenshot${complete.length}.png`});
-    console.log(complete);
+    await page.waitForSelector('h1').then(() => complete.push(Date.now()-complete.pop()));
+    console.log(complete.slice(-1) + 'ms: ' + page._target._targetInfo.url);
 
     // convert events to HAR file
     // const har = harFromMessages(events);
